backend/requirements.txt

fastapi
uvicorn
beanie
motor
pymongo
python-dotenv
python-jose[cryptography]
pyjwt
requests
python-multipart

fastapi==0.104.1
uvicorn[standard]==0.24.0
beanie==1.23.1
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
pyjwt==2.8.0
cryptography==41.0.7
requests==2.31.0
python-multipart==0.0.6
splitwise==3.0.0